Move websys-as-gles bindings to separate crate
The main motivation is to separate the necessary web-sys features to
make it easier to integrate in other places in the future. We also might
want to check out the glow crate. However, they seem to maintain their
bindings manually. So contributing webgl_gen_websys2gles upstream later
may improve the ecosystem...