Write some audio code
Current idea is to return from the audio callback as quickly as possible
by memcopying into one of 2 vecs (allocations) which are rotating
through a 2-element-wide bounded channel. This allows both the producing
side as well as the receiving side to take ownership of a location at
any given time. At the same time, the channel guarantees that the
producer always overwrites the oldest data (ordering guarantee). The
drawback is that the receiver may only get the second-newest sample
buffer. I'll probably need to think about this a bit more...